pyproject.toml
[project]
name = "caixa-nfse"
version = "0.1.0"
description = "Sistema de Caixa com NFS-e - Compliance Contábil/Fiscal"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Proprietary"}
authors = [
    {name = "Rodrigo", email = "dev@example.com"}
]

dependencies = [
    # Django Stack
    "django>=5.0,<6.0",
    "djangorestframework>=3.15",
    "django-filter>=24.0",
    "django-tables2>=2.7",
    "django-crispy-forms>=2.3",
    "crispy-bootstrap5>=2024.2",
    
    # Database
    "psycopg[binary]>=3.2",
    "django-encrypted-model-fields>=0.6",
    
    # Async Tasks
    "celery[redis]>=5.4",
    "redis>=5.0",
    
    # NFS-e / XML
    "lxml>=5.2",
    "signxml>=4.0",
    "zeep>=4.2",
    "cryptography>=43.0",
    
    # Observability
    "sentry-sdk[django,celery]>=2.0",
    "opentelemetry-sdk>=1.25",
    "opentelemetry-instrumentation-django>=0.46b0",
    "opentelemetry-instrumentation-celery>=0.46b0",
    
    # Utils
    "python-decouple>=3.8",
    "gunicorn>=22.0",
    "whitenoise>=6.7",
    "django-htmx>=1.18",
    "django-widget-tweaks>=1.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2",
    "pytest-django>=4.8",
    "pytest-cov>=5.0",
    "pytest-asyncio>=0.23",
    "factory-boy>=3.3",
    "faker>=26.0",
    "black>=24.4",
    "ruff>=0.5",
    "mypy>=1.10",
    "django-stubs>=5.0",
    "djangorestframework-stubs>=3.15",
    "pre-commit>=3.7",
    "ipython>=8.26",
    "django-debug-toolbar>=4.4",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["caixa_nfse*"]

[tool.black]
line-length = 100
target-version = ["py312"]
exclude = '''
/(
    \.git
    | \.mypy_cache
    | \.pytest_cache
    | \.venv
    | migrations
)/
'''

[tool.ruff]
line-length = 100
target-version = "py312"
exclude = ["migrations"]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "DJ",     # flake8-django
]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["caixa_nfse"]

[tool.mypy]
python_version = "3.12"
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
strict = true
warn_unreachable = true
ignore_missing_imports = true

[tool.django-stubs]
django_settings_module = "caixa_nfse.settings.local"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "caixa_nfse.settings.test"
python_files = ["test_*.py"]
addopts = "-v --tb=short"
